Fix transmission by changing sent type to byte[]
Communication between the client and server calculates the length of
the message incorrectly. This is due to the fact that this was calcula-
ted from the length of the String to be sent. Changing the String to
byte[] before sending solves the problem.

Patch also adds UTF-8 character encoding to the sent messages in
client.